+++ Source/AutoPriorities/Tests/Tests.cs
@@ -12,7 +12,7 @@ namespace Tests
         [Test]
         public void TestIterPercents()
         {
-            var percents = new List<float> {0.1f, 0.2f, 0.3f};
+            var percents = new List<double> {0.1, 0.2, 0.3};
 
             Assert.AreEqual(percents.IterPercents(10).ToArray(),
                 new[] {(0, 0), (1, 1), (2, 1), (3, 2), (4, 2), (5, 2)});
